Refactor branding files
